organized components into individual folders
