+++ src/components/loginForm/LoginForm.js
@@ -1,6 +1,6 @@
 import React from "react";
 import Spinner from "react-spinkit";
-import { withAsyncAction } from "../HOCs";
+import { withAsyncAction } from "../../HOCs";
 import "./LoginForm.css";
 
 class LoginForm extends React.Component {
