+++ src/pages/Profile.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Menu from "../components/Menu";
+import Menu from "../components/menu/Menu";
 import { userIsAuthenticated } from "../HOCs";
 
 class Profile extends React.Component {
